Revert "66 set up instructor and practitioner view page (#85)"
This reverts commit 58623bdce6396cf670d537f1e423bec752318de4.
