+++ src/back/server.py
@@ -8,16 +8,12 @@ import uuid
 
 import cv2
 from aiohttp import web
-from aiortc import (
-    MediaStreamTrack,
-    RTCPeerConnection,
-    RTCSessionDescription,
-    VideoStreamTrack,
-)
+from aiortc import MediaStreamTrack, RTCPeerConnection, RTCSessionDescription, VideoStreamTrack
 from aiortc.contrib.media import MediaRelay
 from av import VideoFrame
 
 import mediapipe as mp
+
 
 ROOT = os.path.dirname(__file__)
 
@@ -25,22 +21,18 @@ logger = logging.getLogger("pc")
 pcs = set()
 relay = MediaRelay()
 consumer_track = VideoStreamTrack()
-effects = dict()
+annotation = None
 
 mp_pose = mp.solutions.pose
-pose = mp_pose.Pose(
-    static_image_mode=False,
-    model_complexity=1,
-    enable_segmentation=False,
-    smooth_landmarks=True,
-)
+pose = mp_pose.Pose(static_image_mode=False, model_complexity=1, enable_segmentation=False, smooth_landmarks=True)
 # Drawing utility
 mp_drawing = mp.solutions.drawing_utils
 
 
+
 class VideoTransformTrack(MediaStreamTrack):
     """
-    A video stream track that transforms frames from another track.
+    A video stream track that transforms frames from an another track.
     """
 
     kind = "video"
@@ -119,21 +111,19 @@ class VideoTransformTrack(MediaStreamTrack):
         else:
             return frame
 
-
 async def consumer(request):
     if request.method == "OPTIONS":
         return web.Response(
             content_type="application/json",
-            headers={"Access-Control-Allow-Origin": "*",
-                     "Access-Control-Allow-Credentials": "true",
-                     "Access-Control-Allow-Methods": "POST, GET, OPTIONS",
-                     "Access-Control-Allow-Headers": "Content-Type"},
+            headers={"Access-Control-Allow-Origin": "*", 
+                        "Access-Control-Allow-Credentials": "true", 
+                        "Access-Control-Allow-Methods": "POST, GET, OPTIONS", 
+                        "Access-Control-Allow-Headers": "Content-Type"},
         )
 
     params = await request.json()
-    annotation = params["video_transform"]
-
     description = RTCSessionDescription(sdp=params["sdp"], type=params["type"])
+
     pc = RTCPeerConnection()
     pc_id = "PeerConnection(%s)" % uuid.uuid4()
     pcs.add(pc)
@@ -143,7 +133,8 @@ async def consumer(request):
 
     log_info("Track %s sent", consumer_track.kind)
     pc.addTrack(
-        relay.subscribe(effects[annotation])
+        # VideoTransformTrack(relay.subscribe(consumer_track), transform=params["video_transform"])
+        VideoTransformTrack(relay.subscribe(consumer_track), transform=annotation)
     )
 
     await pc.setRemoteDescription(description)
@@ -155,10 +146,10 @@ async def consumer(request):
         text=json.dumps(
             {"sdp": pc.localDescription.sdp, "type": pc.localDescription.type}
         ),
-        headers={"Access-Control-Allow-Origin": "*",
-                 "Access-Control-Allow-Credentials": "true",
-                 "Access-Control-Allow-Methods": "POST, GET, OPTIONS",
-                 "Access-Control-Allow-Headers": "Content-Type"},
+        headers={"Access-Control-Allow-Origin": "*", 
+            "Access-Control-Allow-Credentials": "true", 
+            "Access-Control-Allow-Methods": "POST, GET, OPTIONS", 
+            "Access-Control-Allow-Headers": "Content-Type"},
     )
 
 
@@ -166,14 +157,16 @@ async def broadcast(request):
     if request.method == "OPTIONS":
         return web.Response(
             content_type="application/json",
-            headers={"Access-Control-Allow-Origin": "*",
-                     "Access-Control-Allow-Credentials": "true",
-                     "Access-Control-Allow-Methods": "POST, GET, OPTIONS",
-                     "Access-Control-Allow-Headers": "Content-Type"},
+            headers={"Access-Control-Allow-Origin": "*", 
+                        "Access-Control-Allow-Credentials": "true", 
+                        "Access-Control-Allow-Methods": "POST, GET, OPTIONS", 
+                        "Access-Control-Allow-Headers": "Content-Type"},
         )
 
     params = await request.json()
     offer = RTCSessionDescription(sdp=params["sdp"], type=params["type"])
+    global annotation
+    annotation=params["video_transform"]
 
     pc = RTCPeerConnection()
     pc_id = "PeerConnection(%s)" % uuid.uuid4()
@@ -205,11 +198,6 @@ async def broadcast(request):
         if track.kind == "video":
             global consumer_track
             consumer_track = track
-            effects["source"] = relay.subscribe(track)
-            effects["skeleton"] = VideoTransformTrack(relay.subscribe(track), "skeleton")
-            effects["cartoon"] = VideoTransformTrack(relay.subscribe(track), "cartoon")
-            effects["edges"] = VideoTransformTrack(relay.subscribe(track), "edges")
-            effects["rotate"] = VideoTransformTrack(relay.subscribe(track), "rotate")
 
         @track.on("ended")
         async def on_ended():
@@ -227,10 +215,10 @@ async def broadcast(request):
         text=json.dumps(
             {"sdp": pc.localDescription.sdp, "type": pc.localDescription.type}
         ),
-        headers={"Access-Control-Allow-Origin": "*",
-                 "Access-Control-Allow-Credentials": "true",
-                 "Access-Control-Allow-Methods": "POST, GET, OPTIONS",
-                 "Access-Control-Allow-Headers": "Content-Type"},
+        headers={"Access-Control-Allow-Origin": "*", 
+            "Access-Control-Allow-Credentials": "true", 
+            "Access-Control-Allow-Methods": "POST, GET, OPTIONS", 
+            "Access-Control-Allow-Headers": "Content-Type"},
     )
 
 
@@ -251,11 +239,10 @@ async def process_frame_for_skeleton(frame):
     # Draw the pose annotations on the frame
     annotated_frame = frame.copy()
     if results.pose_landmarks:
-        mp_drawing.draw_landmarks(
-            annotated_frame, results.pose_landmarks, mp_pose.POSE_CONNECTIONS
-        )
+        mp_drawing.draw_landmarks(annotated_frame, results.pose_landmarks, mp_pose.POSE_CONNECTIONS)
 
     return annotated_frame
+
 
 
 if __name__ == "__main__":
