+++ src/client/src/pages/Home.tsx
@@ -1,0 +1,5 @@
+function Home() {
+  return <h1>Hello World</h1>;
+}
+
+export default Home;
